Refactor System.Diagnostics.Process with a PAL layer
