+++ src/System.Diagnostics.Process/src/Microsoft/Win32/SafeHandles/SafeProcessHandle.cs
@@ -13,24 +13,28 @@
 using System;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
+using System.Security;
 
 namespace Microsoft.Win32.SafeHandles
 {
-    [System.Security.SecurityCriticalAttribute]
-    public sealed class SafeProcessHandle : SafeHandle
+    [SecurityCritical]
+    public sealed partial class SafeProcessHandle : SafeHandle
     {
-        internal static SafeProcessHandle InvalidHandle = new SafeProcessHandle(IntPtr.Zero);
+        internal static readonly SafeProcessHandle InvalidHandle = new SafeProcessHandle(new IntPtr(DefaultInvalidHandleValue));
 
-        internal SafeProcessHandle() : base(IntPtr.Zero, true) { }
+        internal SafeProcessHandle()
+            : base(new IntPtr(DefaultInvalidHandleValue), true) 
+        {
+        }
 
         internal SafeProcessHandle(IntPtr handle)
-            : base(IntPtr.Zero, true)
+            : base(new IntPtr(DefaultInvalidHandleValue), true)
         {
             SetHandle(handle);
         }
 
         public SafeProcessHandle(IntPtr handle, bool ownsHandle)
-            : base(IntPtr.Zero, ownsHandle)
+            : base(new IntPtr(DefaultInvalidHandleValue), ownsHandle)
         {
             SetHandle(handle);
         }
@@ -39,19 +43,6 @@ namespace Microsoft.Win32.SafeHandles
         {
             Debug.Assert(IsInvalid, "Safe handle should only be set once");
             base.handle = h;
-        }
-
-        public override bool IsInvalid
-        {
-            [System.Security.SecurityCritical]
-            get
-            { return handle == new IntPtr(0) || handle == new IntPtr(-1); }
-        }
-
-        [System.Security.SecurityCritical]
-        override protected bool ReleaseHandle()
-        {
-            return Interop.mincore.CloseHandle(handle);
         }
     }
 }
