+++ src/System.Diagnostics.Process/src/System.Diagnostics.Process.csproj
@@ -17,7 +17,9 @@
   </PropertyGroup>
   <!-- References are resolved from packages.config -->
   <ItemGroup>
-    <Compile Include="Interop\Interop.Manual.cs" />
+    <Compile Include="$(CommonPath)\System\SR.cs">
+      <Link>Common\System\SR.cs</Link>
+    </Compile>
     <Compile Include="Microsoft\Win32\SafeHandles\SafeProcessHandle.cs" />
     <Compile Include="Microsoft\Win32\SafeHandles\SafeThreadHandle.cs" />
     <Compile Include="Microsoft\Win32\SafeHandles\SafeTokenHandle.cs" />
@@ -29,8 +31,9 @@
     </Compile>
     <Compile Include="System\Diagnostics\AsyncStreamReader.cs" />
     <Compile Include="System\Diagnostics\DataReceivedEventArgs.cs" />
-    <Compile Include="System\Diagnostics\PerformanceCounterLib.cs" />
+    <Compile Include="System\Diagnostics\ModuleInfo.cs" />
     <Compile Include="System\Diagnostics\Process.cs" />
+    <Compile Include="System\Diagnostics\ProcessInfo.cs" />
     <Compile Include="System\Diagnostics\ProcessManager.cs" />
     <Compile Include="System\Diagnostics\ProcessModule.cs" />
     <Compile Include="System\Diagnostics\ProcessModuleCollection.cs" />
@@ -38,16 +41,36 @@
     <Compile Include="System\Diagnostics\ProcessStartInfo.cs" />
     <Compile Include="System\Diagnostics\ProcessThread.cs" />
     <Compile Include="System\Diagnostics\ProcessThreadCollection.cs" />
-    <Compile Include="System\Diagnostics\ProcessWaitHandle.cs" />
     <Compile Include="System\Diagnostics\ProcessWindowStyle.cs" />
+    <Compile Include="System\Diagnostics\ThreadInfo.cs" />
     <Compile Include="System\Diagnostics\ThreadPriorityLevel.cs" />
     <Compile Include="System\Diagnostics\ThreadState.cs" />
     <Compile Include="System\Diagnostics\ThreadWaitReason.cs" />
   </ItemGroup>
-  <ItemGroup>
-    <Compile Include="$(CommonPath)\System\SR.cs" >
-      <Link>Common\System\SR.cs</Link>
+  <ItemGroup Condition=" '$(OS)' == 'Windows_NT' ">
+    <Compile Include="Interop\Interop.Windows.cs" />
+    <Compile Include="Microsoft\Win32\SafeHandles\SafeProcessHandle.Windows.cs" />
+    <Compile Include="Microsoft\Win32\SafeHandles\SafeThreadHandle.Windows.cs" />
+    <Compile Include="Microsoft\Win32\SafeHandles\SafeTokenHandle.Windows.cs" />
+    <Compile Include="System\Diagnostics\PerformanceCounterLib.cs" />
+    <Compile Include="System\Diagnostics\Process.Windows.cs" />
+    <Compile Include="System\Diagnostics\ProcessManager.Windows.cs" />
+    <Compile Include="System\Diagnostics\ProcessThread.Windows.cs" />
+    <Compile Include="System\Diagnostics\ProcessThreadTimes.cs" />
+    <Compile Include="System\Diagnostics\ProcessWaitHandle.Windows.cs" />
+  </ItemGroup>
+  <ItemGroup Condition=" '$(OS)' == 'Unix' ">
+    <Compile Include="$(CommonPath)\System\NotImplemented.cs">
+      <!-- TODO: Remove once implemented -->
+      <Link>Common\System\NotImplemented.cs</Link>
     </Compile>
+    <Compile Include="Microsoft\Win32\SafeHandles\SafeProcessHandle.Unix.cs" />
+    <Compile Include="Microsoft\Win32\SafeHandles\SafeThreadHandle.Unix.cs" />
+    <Compile Include="Microsoft\Win32\SafeHandles\SafeTokenHandle.Unix.cs" />
+    <Compile Include="System\Diagnostics\Process.Unix.cs" />
+    <Compile Include="System\Diagnostics\ProcessManager.Unix.cs" />
+    <Compile Include="System\Diagnostics\ProcessThread.Unix.cs" />
+    <Compile Include="System\Diagnostics\ProcessWaitHandle.Unix.cs" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Resources\Strings.resx">
