+++ src/System.Diagnostics.Process/src/System/Diagnostics/DataReceivedEventArgs.cs
@@ -1,27 +1,22 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
-
 namespace System.Diagnostics
 {
-    public delegate void DataReceivedEventHandler(Object sender, DataReceivedEventArgs e);
+    public delegate void DataReceivedEventHandler(object sender, DataReceivedEventArgs e);
 
     public class DataReceivedEventArgs : EventArgs
     {
-        internal String _data;
+        internal readonly string _data;
 
-        internal DataReceivedEventArgs(String data)
+        internal DataReceivedEventArgs(string data)
         {
             _data = data;
         }
 
-        public String Data
+        public string Data
         {
-            get
-            {
-                return _data;
-            }
+            get { return _data; }
         }
     }
 }
