+++ src/System.Diagnostics.Process/src/System/Diagnostics/ProcessModule.cs
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
 using System.Globalization;
 
 namespace System.Diagnostics
