+++ src/System.Diagnostics.Process/src/System/Diagnostics/ProcessModuleCollection.cs
@@ -1,10 +1,8 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Diagnostics;
 
 namespace System.Diagnostics
 {
