+++ src/System.Diagnostics.Process/src/System/Diagnostics/ProcessThreadCollection.cs
@@ -3,10 +3,6 @@
 
 using System.Collections;
 using System.Collections.Generic;
-using System;
-using System.IO;
-using System.ComponentModel;
-using System.Diagnostics;
 
 namespace System.Diagnostics
 {
