+++ src/System.Diagnostics.Process/src/System/Diagnostics/ProcessWaitHandle.Unix.cs
@@ -1,0 +1,17 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using Microsoft.Win32.SafeHandles;
+using System.Runtime.InteropServices;
+using System.Threading;
+
+namespace System.Diagnostics
+{
+    internal sealed class ProcessWaitHandle : WaitHandle
+    {
+        internal ProcessWaitHandle(SafeProcessHandle processHandle)
+        {
+            throw NotImplemented.ByDesign; // TODO: Implement this
+        }
+    }
+}
