+++ src/System.Diagnostics.Process/src/System/Diagnostics/ProcessWaitHandle.Windows.cs
@@ -1,17 +1,15 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using Microsoft.Win32;
 using Microsoft.Win32.SafeHandles;
 using System.Runtime.InteropServices;
 using System.Threading;
 
 namespace System.Diagnostics
 {
-    internal class ProcessWaitHandle : WaitHandle
+    internal sealed class ProcessWaitHandle : WaitHandle
     {
         internal ProcessWaitHandle(SafeProcessHandle processHandle)
-            : base()
         {
             SafeWaitHandle waitHandle = null;
             SafeProcessHandle currentProcHandle = Interop.mincore.GetCurrentProcess();
