+++ src/System.Diagnostics.Process/src/System/Diagnostics/ThreadPriorityLevel.cs
@@ -1,10 +1,6 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System.Threading;
-
-using System.Diagnostics;
-
 namespace System.Diagnostics
 {
     /// <devdoc>
